Module Decoder.File

Local file decoding.

Convenience module to * decode local files

Types

type handle = {

fd : Unix.file_descr;

dec : t;

info : info;

comments : (string * (string * string) list) option;

}

Functions

val create_from_fd : write:(float array array -> unit) -> Unix.file_descr -> handle

Create a file decoder from a Unix file * descriptor * * Note: this decoder requires seeking thus will only work on seekable * file descriptor.

val create : write:(float array array -> unit) -> string -> handle

Create a file decoder from a file URI