Module Encoder.File

Encode to a local file

Convenience module to encode to a local native FLAC file.

Types

type handle = {

fd : Unix.file_descr;

enc : t;

}

Handle for file encoder

Functions

val create_from_fd : ?comments:comments -> params -> Unix.file_descr -> handle

Create a file encoder writing data to a given Unix file descriptor. * * Note: this encoder requires seeking thus will only work on seekable * file descriptor.

val create : ?comments:comments -> params -> string -> handle

Create a file encoder writing data to the given file URI